==> python/0 Merge Class Datasets.ipynb <==
import pandas as pd, boto3, re, os
pd.set_option("display.max_rows",1000)
pd.set_option("display.max_colwidth",1000)

public_classes = pd.read_csv("docs_module/metadata/class_metadata.csv")
licensed_classes = pd.read_csv("docs_module/metadata/class_metadata_licensed.csv")

class_metadata = pd.concat([licensed_classes, public_classes], sort=False).reset_index(drop=True)

array_to_text = ["inputs","output","tags"]
for c in array_to_text:
    class_metadata[c] = class_metadata[c].str.replace("[\[\]'‘]","")

class_metadata.to_csv("docs_module/metadata/class_metadata_all.csv", index=False)
